+++ src/index.ts
@@ -34,6 +34,10 @@ export {IProps as CheckboxProps} from './modules/checkbox/checkbox.types';
 export {default as Chip} from './modules/chip/chip';
 export {IProps as ChipProps} from './modules/chip/chip.types';
 
+// Compact Card List Item
+export {default as CompactCardListItem} from './modules/compact-card-list/compact-card-list-item';
+export {IProps as CompactCardListItemProps} from './modules/compact-card-list/compact-card-list-item.types';
+
 // Date Picker
 export {default as DataBlock} from './modules/data-block/data-block';
 export {IProps as DataBlockProps} from './modules/data-block/data-block.types';
@@ -69,6 +73,10 @@ export {IProps as IconProps} from './modules/icon/icon-external.types';
 // Icon Toggle
 export {default as IconToggle} from './modules/icon-toggle/icon-toggle';
 export {IProps as IconToggleProps} from './modules/icon-toggle/icon-toggle.types';
+
+// Linear Gradient View
+export {default as LinearGradientView} from './modules/linear-gradient/linear-gradient-view';
+export {IProps as LinearGradientViewProps} from './modules/linear-gradient/linear-gradient-view.types';
 
 // Link
 export {default as Link} from './modules/link/link';
@@ -116,6 +124,10 @@ export {
   IListItemProps as SelectableListItemProps,
 } from './modules/selectable-list/selectable-list.types';
 
+// Tabs
+export {default as Tabs} from './modules/tabs/tabs';
+export {IProps as TabsProps} from './modules/tabs/tabs.types';
+
 // Text
 export {default as Text} from './modules/text/text';
 export {IProps as TextProps} from './modules/text/text.types';
@@ -138,3 +150,7 @@ export {IProps as TilesProps} from './modules/tiles/tiles.types';
 
 // Utilities
 export {DeviceUtils, LayoutUtils} from './modules/utilities';
+
+// People-list-item
+export {default as PeopleListItem} from './modules/people-list/people-list-item';
+export {IProps as IPeopleItemProps} from './modules/people-list/people-list-item.types';